Naming in Notebook-Add, made url endpoint /notebooks for dashboard
