+++ src/components/notebooks/cards/notebook-add.tsx
@@ -1,5 +1,5 @@
 
-export default function NotebookPlusCard() {
+export default function NotebookAddCard() {
   return (
     <a href="#/" className="rounded-notebook-card justify-center min-w-[120px] max-w-[380px] bg-component-faded">
       <h1 className="text-5xl text-center self-center text-faded" >+</h1>
